refactor(sleep): extract a shared SleepModel.build() out of SleepView
Pure reorganization, no behavior change. The Sleep-tab model was built by
SleepView.buildModel() from ~15 private instance computeds + the decoded-Night
pipeline, so no other host could build it. Move the value types (Stages, Night,
SleepModel) and the derivation pipeline (decodedNight, navDays, mainNightGroup
reuse, nap crediting, typical means, the 7 per-tile series, debt ledger, trend
points) into a new Strand/Screens/SleepModel.swift as a pure
SleepModel.build(SleepModelInputs) - a function only of its inputs (days,
sleeps, allSessions, importedSleep, habitualMidsleepSec, motionByStart).

SleepView.buildModel() is now a thin wrapper snapshotting its state into the
inputs; navDays/decodedNight became thin wrappers over the shared statics (one
source of truth, no duplicated math). Every renderer + renderer-only helper
stays on SleepView unchanged. No cards, no HostedCard/Today wiring, no Android -
this only unblocks the Today host building the same model for the hosted sleep
cards (byte-identical to the Sleep tab by construction).
